requirements.txt
openai
python-dotenv
pinecone-client
pinecone-datasets
langchain
datasets
icecream
torch==2.2.1
tiktoken
fastapi 
httpx
uvicorn
websockets